replace *_as_string_or_struct with polymorphic impl;
