+++ ergo-lib/src/chain/transaction/input/prover_result/json.rs
@@ -1,15 +1,8 @@
-use std::fmt;
-use std::marker::PhantomData;
 use std::str::FromStr;
 
 use crate::chain::json::context_extension::ContextExtensionSerde;
 use ergotree_interpreter::sigma_protocol::prover::ContextExtension;
-use serde::de;
-use serde::de::MapAccess;
-use serde::de::Visitor;
 use serde::ser::SerializeStruct;
-use serde::Deserialize;
-use serde::Deserializer;
 use serde::Serialize;
 
 use super::ProverResult;
@@ -39,56 +32,6 @@ impl FromStr for ProverResult {
             extension: ContextExtension::empty(),
         })
     }
-}
-
-// via https://serde.rs/string-or-struct.html
-pub fn proof_as_string_or_struct<'de, T, D>(deserializer: D) -> Result<T, D::Error>
-where
-    T: Deserialize<'de> + FromStr<Err = base16::DecodeError>,
-    D: Deserializer<'de>,
-{
-    // This is a Visitor that forwards string types to T's `FromStr` impl and
-    // forwards map types to T's `Deserialize` impl. The `PhantomData` is to
-    // keep the compiler from complaining about T being an unused generic type
-    // parameter. We need T in order to know the Value type for the Visitor
-    // impl.
-    struct StringOrStruct<T>(PhantomData<fn() -> T>);
-
-    impl<'de, T> Visitor<'de> for StringOrStruct<T>
-    where
-        T: Deserialize<'de> + FromStr<Err = base16::DecodeError>,
-    {
-        type Value = T;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-            formatter.write_str("string or map")
-        }
-
-        fn visit_str<E>(self, value: &str) -> Result<T, E>
-        where
-            E: de::Error,
-        {
-            FromStr::from_str(value).map_err(|e| {
-                de::Error::custom(format!(
-                    "error: {}, while parsing proof bytes from string: {:?}",
-                    e, value
-                ))
-            })
-        }
-
-        fn visit_map<M>(self, map: M) -> Result<T, M::Error>
-        where
-            M: MapAccess<'de>,
-        {
-            // `MapAccessDeserializer` is a wrapper that turns a `MapAccess`
-            // into a `Deserializer`, allowing it to be used as the input to T's
-            // `Deserialize` implementation. T then deserializes itself using
-            // the entries from the map visitor.
-            Deserialize::deserialize(de::value::MapAccessDeserializer::new(map))
-        }
-    }
-
-    deserializer.deserialize_any(StringOrStruct(PhantomData))
 }
 
 #[allow(clippy::unwrap_used)]
