+++ ergotree-ir/src/chain/json.rs
@@ -1,5 +1,14 @@
 //! JSON serialization
 
+use std::fmt;
+use std::marker::PhantomData;
+use std::str::FromStr;
+
+use serde::de;
+use serde::de::MapAccess;
+use serde::de::Visitor;
+use serde::Deserialize;
+use serde::Deserializer;
 use serde::Serializer;
 
 pub(crate) mod box_value;
@@ -15,4 +24,57 @@ where
     T: AsRef<[u8]>,
 {
     serializer.serialize_str(&base16::encode_lower(bytes.as_ref()))
+}
+
+// via https://serde.rs/string-or-struct.html
+/// Parses type either from a string or a struct
+pub fn t_as_string_or_struct<'de, T, D, E>(deserializer: D) -> Result<T, D::Error>
+where
+    T: Deserialize<'de> + FromStr<Err = E>,
+    D: Deserializer<'de>,
+    E: std::fmt::Display,
+{
+    // This is a Visitor that forwards string types to T's `FromStr` impl and
+    // forwards map types to T's `Deserialize` impl. The `PhantomData` is to
+    // keep the compiler from complaining about T being an unused generic type
+    // parameter. We need T in order to know the Value type for the Visitor
+    // impl.
+    struct StringOrStruct<T>(PhantomData<fn() -> T>);
+
+    impl<'de, T, FromStrErr> Visitor<'de> for StringOrStruct<T>
+    where
+        T: Deserialize<'de> + FromStr<Err = FromStrErr>,
+        FromStrErr: std::fmt::Display,
+    {
+        type Value = T;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("string or map")
+        }
+
+        fn visit_str<E>(self, value: &str) -> Result<T, E>
+        where
+            E: de::Error,
+        {
+            FromStr::from_str(value).map_err(|e| {
+                de::Error::custom(format!(
+                    "error: {}, while parsing from string: {:?}",
+                    e, value
+                ))
+            })
+        }
+
+        fn visit_map<M>(self, map: M) -> Result<T, M::Error>
+        where
+            M: MapAccess<'de>,
+        {
+            // `MapAccessDeserializer` is a wrapper that turns a `MapAccess`
+            // into a `Deserializer`, allowing it to be used as the input to T's
+            // `Deserialize` implementation. T then deserializes itself using
+            // the entries from the map visitor.
+            Deserialize::deserialize(de::value::MapAccessDeserializer::new(map))
+        }
+    }
+
+    deserializer.deserialize_any(StringOrStruct(PhantomData))
 }
